requirements-dev.txt
--extra-index-url https://backend:dc222fa6-452f-4253-a2fd-a52e7ba2e7e9@nexus.sigame.com.br/repository/pypi/simple
aiohttp==3.8.1
aioredis==2.0.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
etria-logger==0.1.9
frozenlist==1.3.0
idna==3.3
multidict==6.0.2
nidavellir==0.1.0
orjson==3.6.8
python-decouple==3.6
python-json-logger==2.0.2
typing-extensions==4.1.1
yarl==1.7.2

pytest==6.2.5
pytest-asyncio==0.18.3
pip-chill==1.0.1